Move dependency finding code to DependencyFinder
