+++ core/src/main/java/com/nikodoko/javaimports/resolver/maven/MavenDependencyFinder.java
@@ -1,7 +1,11 @@
 package com.nikodoko.javaimports.resolver.maven;
 
+import java.io.FileReader;
 import java.io.IOException;
 import java.io.Reader;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.maven.model.Dependency;
@@ -39,5 +43,33 @@ class MavenDependencyFinder {
 
   List<MavenDependency> result() {
     return dependencies;
+  }
+
+  List<MavenDependency> findAll(Path moduleRoot) throws IOException {
+    scanPom(moduleRoot);
+
+    Path target = moduleRoot.getParent();
+    while (target != null && !allFound) {
+      if (!hasPom(target)) {
+        break;
+      }
+
+      scanPom(target);
+      target = target.getParent();
+    }
+
+    return dependencies;
+  }
+
+  private boolean hasPom(Path directory) {
+    Path pom = Paths.get(directory.toString(), "pom.xml");
+    return Files.exists(pom);
+  }
+
+  private void scanPom(Path directory) throws IOException {
+    Path pom = Paths.get(directory.toString(), "pom.xml");
+    try (FileReader reader = new FileReader(pom.toFile())) {
+      scan(reader);
+    }
   }
 }
